driver: move XDMA channel init function to xdma_sgdma.c
We need to preserve this function to be later used by the Hermes driver.
Everything else on xdma_cdev.* is no longer required and will be removed
in another patch.
